+++ Tests/ChessCoreTests/Rules/Standard/StandardRulesTests.swift
@@ -49,7 +49,7 @@ import Testing
         ("8/8/8/4N3/8/8/8/8 w - - 0 1", "c6 d7 f7 g6 g4 f3 d3 c4", "e5"),  // Center board.
         ("7N/8/8/8/8/8/8/8 w - - 0 1", "f7 g6", "h8"),  // Corner.
         ("8/5P2/2K5/4N3/8/3R4/8/8 w - - 0 1", "d7 g6 g4 f3 c4", "e5"),  // Friendly blockers.
-        ("8/5p2/2k5/4N3/8/3r4/8/8 w - - 0 1", "c6 d7 f7 g6 g4 f3 d3 c4", "e5"),  // Captures.
+        ("8/5p2/2k5/4N3/8/3r4/8/8 w - - 0 1", "d7 f7 g6 g4 f3 d3 c4", "e5"),  // Captures, excluding the king.
 
         // King moves.
         ("8/8/8/8/4K3/8/8/8 w - - 0 1", "d5 e5 f5 d4 f4 d3 e3 f3", "e4"),  // Center board.